+++ scripts/background.js
@@ -52,7 +52,6 @@ function ccMenus() {
 			chrome.contextMenus.removeAll();
 			chrome.storage.local.get(null, function(options) {
 				if(options['s_type'] == "PHP") {
-					
 					if(options.actions.crsrv === true) {
 						chrome.commands.getAll((commands) => {
 							for (let {name, shortcut} of commands) {
@@ -105,7 +104,7 @@ function bookmarkTab() {
 		chrome.storage.local.get(null, function(options) {
 			if(options.actions.create === false) {				
 				let jsonMark = encodeURIComponent(JSON.stringify({ 
-					"id": Math.random().toString(24).substr(2, 12),
+					"id": Math.random().toString(24).substring(2, 12),
 					"url": tabs[0].url,
 					"title": tabs[0].title,
 					"type": 'bookmark',
@@ -115,18 +114,35 @@ function bookmarkTab() {
 				}));
 				
 				let data = "client=bookmarkTab&caction=addmark&bookmark=" + jsonMark + "&s=false";
-
 				var xhr = new XMLHttpRequest();
 				xhr.open("POST", options['wdurl'], true);
-				xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+				let tarr = {};
+				tarr['client'] = options['s_uuid'];
+				tarr['token'] = options['token'];
+				if(tarr['token'] == '') return false;
+				xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 				xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 				xhr.withCredentials = true;
-				xhr.onload = function () {
+				xhr.onload = async function () {
 					if( xhr.status < 200 || xhr.status > 226) {
 						notify('error', xhr.response);
 						loglines = logit("Error: " + xhr.response);
 					} else {
 						let response = JSON.parse(xhr.response);
+						let xtResponse = xhr.getResponseHeader("X-Request-Info");
+						if(xtResponse !== null) {
+							if(xtResponse !== '0') {
+								await chrome.storage.local.set({token: xtResponse});
+							} else {
+								chrome.storage.local.set({token: ''});
+								let message = chrome.i18n.getMessage("optionsLoginError");
+								notify('error', message);
+								chrome.browserAction.setBadgeText({text: '!'});
+								chrome.browserAction.setBadgeTextColor({color: "white"});
+								chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+							}
+						}
+
 						response = (response == "1") ? "'"+tabs[0].title+"' added":JSON.parse(xhr.response);
 						notify('info', response);
 						loglines = logit("Info: " + response);
@@ -145,17 +161,34 @@ function sendTab(element) {
 			var cdata = "client=" + options['s_uuid'] + "&caction=getpurl&url=" + encodeURIComponent(url) + "&tg=" + element.target.id;
 			var xhr = new XMLHttpRequest();
 			xhr.open("POST", options['wdurl'], true);
-			xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+			let tarr = {};
+			tarr['client'] = options['s_uuid'];
+			tarr['token'] = options['token'];
+			if(tarr['token'] == '') return false;
+			xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 			xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 			xhr.withCredentials = true;
-			xhr.onload = function () {
+			xhr.onload = async function () {
 				if( xhr.status < 200 || xhr.status > 226) {
 					var message = chrome.i18n.getMessage("sendLinkNot");
 					notify('error',message);
 					loglines = logit('Error: ' + message);
-				}
-				else {
+				} else {
 					loglines = logit("Info: " + chrome.i18n.getMessage("sendLinkYes"));
+					let xtResponse = xhr.getResponseHeader("X-Request-Info");
+
+					if(xtResponse !== null) {
+						if(xtResponse !== '0') {
+							await chrome.storage.local.set({token: xtResponse});
+						} else {
+							chrome.storage.local.set({token: ''});
+							let message = chrome.i18n.getMessage("optionsLoginError");
+							notify('error', message);
+							chrome.browserAction.setBadgeText({text: '!'});
+							chrome.browserAction.setBadgeTextColor({color: "white"});
+							chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+						}
+					}
 				}
 			}
 			loglines = logit("Info: " + chrome.i18n.getMessage("sendLinkYes") + ", Client: " + options['s_uuid']);
@@ -199,7 +232,7 @@ function init() {
 	});
 	get_oMarks();
 	chrome.storage.local.set({last_message: ""});
-	chrome.storage.local.get(null, function(options) {
+	chrome.storage.local.get(null, async function(options) {
 		if(options['wdurl'] === undefined) return false;
 		if(options.actions.crsrv === true) {
 			chrome.commands.getAll((commands) => {
@@ -219,19 +252,32 @@ function init() {
 		let s_startup = options['actions']['startup'] || false;
 		let s_type = options['s_type'] || "";
 
+		if(s_type == 'PHP') {
+			if(options['token'] == '') {
+				chrome.browserAction.setBadgeText({text: '!'});
+				chrome.browserAction.setBadgeTextColor({color: "white"});
+				chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+			} else {
+				chrome.browserAction.setBadgeText({text: ''});
+			}
+		} else if(s_type == 'PHP') {
+			if(options['creds'] == '') {
+				chrome.browserAction.setBadgeText({text: '!'});
+				chrome.browserAction.setBadgeTextColor({color: "white"});
+				chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+			} else {
+				chrome.browserAction.setBadgeText({text: ''});
+			}
+		}
+
 		if( s_startup === true && s_type.indexOf('PHP') == -1) {
 			loglines = logit("Info: Initiate WebDAV startup sync");
 			if(options['wdurl']) getDAVMarks();
 		} else if(s_type.indexOf('PHP') == 0) {
-			if(s_startup === true) {
-				loglines = logit("Info: Initiate startup sync");
-				checkFullSync();
-			}
-
 			if(options['wdurl']) {
-				ccMenus();
-				getClientList();
-				getNotifications();
+				await ccMenus();
+				await getClientList();
+				loglines = logit("Info: Init finished");
 			}
 		}
 	});
@@ -240,19 +286,37 @@ function init() {
 function getNotifications() {
 	loglines = logit("Info: Get notifications for current client.");
 	chrome.storage.local.get(null, function(options) {
+		let data = "client=" + options['s_uuid'] + "&caction=gurls&s=" + options['actions']['startup'];
 		let xhr = new XMLHttpRequest();
-		let data = "client=" + options['s_uuid'] + "&caction=gurls";
 		xhr.open("POST", options['wdurl'], true);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 		
 		xhr.withCredentials = true;
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			if( xhr.status < 200 || xhr.status > 226) {
-				message = "Get list of notifications failed. State: "+xhr.status;
+				message = "Get list of notifications failed. State: " + xhr.status;
 				notify('error',message);
 				loglines = logit('Error: '+message);
 			} else {
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+					} else {
+						chrome.storage.local.set({token: ''});
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+					}
+				}
+
 				if(xhr.responseText.length > 2) {
 					var nData = JSON.parse(xhr.responseText);
 					if(Array.isArray(nData)) {
@@ -268,15 +332,20 @@ function getNotifications() {
 					}
 				}
 			}
+			let s_startup = options['actions']['startup'] || false;
+			if(s_startup === true) {
+				loglines = logit("Info: Start Sync");
+				await checkFullSync();
+			}
 		}
 		xhr.send(data);
 	});
 }
 
 function onTabActivated(tab){
-	pTabs.forEach(function(pTab, index){
+	pTabs.forEach(async function(pTab, index){
 		if(pTab.tID == tab.tabId) {
-			dmNoti(pTab.nID);
+			await dmNoti(pTab.nID);
 			pTabs.splice(index,1);
 		}
 	});
@@ -325,20 +394,37 @@ function getClientList() {
 		let data = "client=" + options['s_uuid'] + "&caction=getclients&s="+options['actions']['startup'];
 		let xhr = new XMLHttpRequest();
 		xhr.open("POST", options['wdurl'], true);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
+
 		xhr.withCredentials = true;
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			if( xhr.status < 200 || xhr.status > 226) {
 				message = "Get list of clients failed. State: "+xhr.status;
 				notify('error',message);
 				loglines = logit('Error: '+message);
 			} else {
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+					} else {
+						chrome.storage.local.set({token: ''});
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+					}
+				}
+
 				cData = JSON.parse(xhr.responseText);
 				
-				chrome.storage.local.set({
-					clist:cData
-				});		
+				chrome.storage.local.set({clist:cData});		
 				
 				chrome.permissions.getAll(function(e) {
 					if(e.permissions.includes('contextMenus')) {
@@ -372,11 +458,11 @@ function getClientList() {
 							});
 							let cnt = cData.length - 1;
 							loglines = logit("Info: List of " + cnt + " clients retrieved successful.");
-						} else {
-							console.info("No clients received");
 						}
 					}
 				});
+
+				getNotifications();
 			}
 		}
 		xhr.send(data);
@@ -439,14 +525,32 @@ function dmNoti(nkey) {
 		let xhr = new XMLHttpRequest();
 		let data = "client=" + options['s_uuid'] + "&caction=durl&durl="+nkey;
 		xhr.open("POST", options['wdurl'], true);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 		xhr.withCredentials = true;
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			if( xhr.status < 200 || xhr.status > 226) {
 				message = "Dismiss notification "+nkey+".";
 				notify('error',message);
 				loglines = logit('Error: '+message);
+			} else {
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+					} else {
+						chrome.storage.local.set({token: ''});
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+					}
+				}
 			}
 		}
 		xhr.send(data);
@@ -485,7 +589,7 @@ function onCreatedCheck(id, bookmark) {
 
 function onMovedCheck(id, bookmark) {
 	get_oMarks();
-	chrome.storage.local.get(null, function(options) {
+	chrome.storage.local.get(null, async function(options) {
 		var s_change = options['actions']['change'] || false;
 		var s_type = options['s_type'] || "";
 		
@@ -493,7 +597,7 @@ function onMovedCheck(id, bookmark) {
 			saveAllMarks();
 		}
 		else if(s_change === true && s_type.indexOf('PHP') == 0) {
-			moveMark(id, bookmark);
+			await moveMark(id, bookmark);
 		}
 	});
 }
@@ -521,17 +625,34 @@ function editMark(eData,id) {
 			let cdata = "client="+s_uuid+"&caction=editmark&bookmark="+jsonMark;
 			var xhr = new XMLHttpRequest();
 			xhr.open("POST", options['wdurl'], true);
-			xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+			let tarr = {};
+			tarr['client'] = options['s_uuid'];
+			tarr['token'] = options['token'];
+			if(tarr['token'] == '') return false;
+			xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 			xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 			xhr.withCredentials = true;
-			xhr.onload = function () {
+			xhr.onload = async function () {
 				if( xhr.status < 200 || xhr.status > 226) {
 					message = chrome.i18n.getMessage("errorEditBookmark") + xhr.status;
 					notify('error', message);
 					chrome.browserAction.setTitle({title: date.toLocaleDateString(undefined,doptions) + ": " + chrome.i18n.getMessage("errorEditBookmark")});
 					loglines = logit('Error: '+message);
-				}
-				else {
+				} else {
+					let xtResponse = xhr.getResponseHeader("X-Request-Info");
+					if(xtResponse !== null) {
+						if(xtResponse !== '0') {
+							await chrome.storage.local.set({token: xtResponse});
+						} else {
+							chrome.storage.local.set({token: ''});
+							let message = chrome.i18n.getMessage("optionsLoginError");
+							notify('error', message);
+							chrome.browserAction.setBadgeText({text: '!'});
+							chrome.browserAction.setBadgeTextColor({color: "white"});
+							chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+						}
+					}
+
 					let response = JSON.parse(xhr.responseText);
 					if(response == 1) {
 							loglines = logit("Info: Bookmark edited successfully at the server");
@@ -557,7 +678,7 @@ function editMark(eData,id) {
 }
 
 function onRemovedCheck(id, bookmark) {
-	chrome.storage.local.get(null, function(options) {
+	chrome.storage.local.get(null, async function(options) {
 		var s_remove = options['actions']['remove'] || false;
 		var s_type = options['s_type'] || "";
 		
@@ -566,8 +687,8 @@ function onRemovedCheck(id, bookmark) {
 			saveAllMarks();
 		}
 		else if(s_remove === true  && s_type.indexOf('PHP') == 0) {
-			delMark(id, bookmark);
-			get_oMarks();
+			await delMark(id, bookmark);
+			await get_oMarks();
 		}
 	});
 }
@@ -584,7 +705,7 @@ function exportPHPMarks(upl=[]) {
 			bookmarks = encodeURIComponent(JSON.stringify(upl));
 			p = 1;
 		}
-		
+
 		chrome.storage.local.get(null, function(options) {
 			if(!("s_uuid" in options)) {
 				var s_uuid = uuidv4();
@@ -597,16 +718,33 @@ function exportPHPMarks(upl=[]) {
 			let cdata = 'client='+s_uuid+'&caction=import&bookmark='+bookmarks+"&p="+p+"&s="+options['actions']['startup'];
 			let xhr = new XMLHttpRequest();
 			xhr.open("POST", options['wdurl'], true);
-			xhr.setRequestHeader('Authorization', 'Basic ' + options['creds']);
+			let tarr = {};
+			tarr['client'] = options['s_uuid'];
+			tarr['token'] = options['token'];
+			if(tarr['token'] == '') return false;
+			xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 			xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 			xhr.withCredentials = true;
-			xhr.onload = function () {
+			xhr.onload = async function () {
 				if( xhr.status < 200 || xhr.status > 226) {
 					message = chrome.i18n.getMessage("errorSaveBookmarks") + xhr.status;
 					notify('error',message);
 					loglines = logit("Error: "+message);
-				}
-				else {
+				} else {
+					let xtResponse = xhr.getResponseHeader("X-Request-Info");
+					if(xtResponse !== null) {
+						if(xtResponse !== '0') {
+							await chrome.storage.local.set({token: xtResponse});
+						} else {
+							chrome.storage.local.set({token: ''});
+							let message = chrome.i18n.getMessage("optionsLoginError");
+							notify('error', message);
+							chrome.browserAction.setBadgeText({text: '!'});
+							chrome.browserAction.setBadgeTextColor({color: "white"});
+							chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+						}
+					}
+
 					let response = JSON.parse(xhr.responseText);
 					if(response == 1) {
 						message = chrome.i18n.getMessage("successExportBookmarks");
@@ -620,7 +758,6 @@ function exportPHPMarks(upl=[]) {
 					}
 				}
 			}
-			//loglines = logit("Info: Sending export of local bookmarks to server");
 			xhr.send(cdata);
 		})
 	});
@@ -641,7 +778,7 @@ function saveAllMarks() {
 			xhr.open("PUT", options['wdurl'] + "/" + filename, true);
 			xhr.withCredentials = true;
 			xhr.setRequestHeader('X-Filename', filename);
-			xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+			xhr.setRequestHeader('Authorization', 'Basic ' + btoa(options['creds']));
 			xhr.onload = function () {
 				if( xhr.status < 200 || xhr.status > 226) {
 					message = chrome.i18n.getMessage("errorSaveBookmarks") + xhr.status;
@@ -677,17 +814,34 @@ function delMark(id, bookmark) {
 		let cdata = "client="+s_uuid+"&caction=delmark&bookmark="+jsonMark;
 		var xhr = new XMLHttpRequest();
 		xhr.open("POST", options['wdurl'], true);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 		xhr.withCredentials = true;
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			if( xhr.status < 200 || xhr.status > 226) {
 				message = chrome.i18n.getMessage("errorRemoveBookmark") + xhr.status;
 				notify('error',message);
 				chrome.browserAction.setTitle({title: date.toLocaleDateString(undefined,doptions) + ": " + chrome.i18n.getMessage("errorRemoveBookmark")});
 				loglines = logit('Error: '+message);
-			}
-			else {
+			} else {
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+					} else {
+						chrome.storage.local.set({token: ''});
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+					}
+				}
+
 				let response = JSON.parse(xhr.responseText);
 				if(response == 1) {
 						loglines = logit("Info: Bookmark removed at the server");
@@ -725,16 +879,33 @@ function moveMark(id, bookmark) {
 				let cdata = "client="+s_uuid+"&caction=movemark&bookmark="+jsonMark+"&s="+options['actions']['startup'];
 				var xhr = new XMLHttpRequest();
 				xhr.open("POST", options['wdurl'], true);
-				xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+				let tarr = {};
+				tarr['client'] = options['s_uuid'];
+				tarr['token'] = options['token'];
+				if(tarr['token'] == '') return false;
+				xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 				xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 				xhr.withCredentials = true;
-				xhr.onload = function () {
+				xhr.onload = async function () {
 					if( xhr.status < 200 || xhr.status > 226) {
 						message = chrome.i18n.getMessage("errorMoveBookmark") + xhr.status;
 						notify('error',message);
 						loglines = logit('Error: '+message);
-					}
-					else {
+					} else {
+						let xtResponse = xhr.getResponseHeader("X-Request-Info");
+						if(xtResponse !== null) {
+							if(xtResponse !== '0') {
+								await chrome.storage.local.set({token: xtResponse});
+							} else {
+								chrome.storage.local.set({token: ''});
+								let message = chrome.i18n.getMessage("optionsLoginError");
+								notify('error', message);
+								chrome.browserAction.setBadgeText({text: '!'});
+								chrome.browserAction.setBadgeTextColor({color: "white"});
+								chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+							}
+						}
+
 						let response = JSON.parse(xhr.responseText);
 						if(response == 1)
 								loglines = logit("Info: Bookmark moved successfully at the server");
@@ -763,7 +934,7 @@ function sendMark(bookmark) {
 		bookmark.type = "bookmark";
 	}
 
-	chrome.bookmarks.get(bookmark.parentId, function(bmark) {
+	chrome.bookmarks.get(bookmark.parentId, async function(bmark) {
 		let jsonMark = encodeURIComponent(JSON.stringify({ 
 			"id": bookmark.id,
 			"url": bookmark.url,
@@ -774,28 +945,43 @@ function sendMark(bookmark) {
 			"added": bookmark.dateAdded
 		}));
 
-		chrome.storage.local.get(null, function(options) {
+		await chrome.storage.local.get(null, function(options) {
 			if(!("s_uuid" in options)) {
 				var s_uuid = uuidv4();
 				chrome.storage.local.set({s_uuid: s_uuid});
 			} else {
 				var s_uuid = options['s_uuid'];
 			}
-
 			let cdata = "client="+s_uuid+"&caction=addmark&bookmark="+jsonMark+"&s="+options['actions']['startup'];
-
 			var xhr = new XMLHttpRequest();
 			xhr.open("POST", options['wdurl'], true);
-			xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+			let tarr = {};
+			tarr['client'] = options['s_uuid'];
+			tarr['token'] = options['token'];
+			if(tarr['token'] == '') return false;
+			xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 			xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 			xhr.withCredentials = true;
-			xhr.onload = function () {
+			xhr.onload = async function () {
 				if( xhr.status < 200 || xhr.status > 226) {
 					message = chrome.i18n.getMessage("errorSaveSingleBookmarks")  + xhr.status;
 					notify('error',message);
 					loglines = logit('Error: '+message);
-				}
-				else {
+				} else {
+					let xtResponse = xhr.getResponseHeader("X-Request-Info");
+					if(xtResponse !== null) {
+						if(xtResponse !== '0') {
+							await chrome.storage.local.set({token: xtResponse});
+						} else {
+							chrome.storage.local.set({token: ''});
+							let message = chrome.i18n.getMessage("optionsLoginError");
+							notify('error', message);
+							chrome.browserAction.setBadgeText({text: '!'});
+							chrome.browserAction.setBadgeTextColor({color: "white"});
+							chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+						}
+					}
+
 					let response = JSON.parse(xhr.responseText);
 					if(response == 1)
 							loglines = logit("Info: Bookmark added successfully at the server");
@@ -824,8 +1010,7 @@ function saveDAVMarks(bookmarkItems) {
 		xhr.open("PUT", options['wdurl'] + "/" + filename, true);
 		xhr.withCredentials = true;
 		xhr.setRequestHeader('X-Filename', filename);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
-		
+		xhr.setRequestHeader('Authorization', 'Basic ' + btoa(options['creds']));
 		xhr.onload = function () {
 			if( xhr.status < 200 || xhr.status > 226) {
 				message = chrome.i18n.getMessage("errorSaveBookmarks") + xhr.status;
@@ -854,10 +1039,14 @@ function getPHPMarks() {
 		let params = 'client='+s_uuid+'&caction=startup&s='+options['actions']['startup'];
 		xhr.open('POST', options['wdurl'] + '?t=' + Math.random(), true);
 		xhr.withCredentials = true;
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 		
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			let datems = Date.now();
 			let date = new Date(datems);
 			let doptions = { weekday: 'short',  hour: '2-digit', minute: '2-digit' };
@@ -867,6 +1056,20 @@ function getPHPMarks() {
 				notify('error',message);
 				loglines = logit('Error: '+message);
 			} else {
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+					} else {
+						chrome.storage.local.set({token: ''});
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+					}
+				}
+
 				response = (xhr.responseText);
 				if(abrowser == false) response = c2cm(response);
 				let PHPMarks = JSON.parse(response);
@@ -891,42 +1094,56 @@ function getPHPMarks() {
 	});
 }
 
-function checkFullSync() {
-	chrome.storage.local.get(null, function(options) {
+async function checkFullSync() {
+	chrome.storage.local.get(null, async function(options) {
 		let xhr = new XMLHttpRequest();
 		let params = 'client=' + options['s_uuid'] + '&caction=cinfo';
-		xhr.open('POST', options['wdurl'] + '?t=' + Math.random(), true);
+		xhr.open('POST', options['wdurl'] + '?t=' + Math.random(), false);
 		xhr.withCredentials = true;
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			if( xhr.status < 200 || xhr.status > 226) {
 				let message = 'FullSync Check failed';
 				notify('error', message);
-				loglines = logit('Error: '+message);
+				loglines = logit('Error: ' + message);
 			} else {
-				let cinfo = JSON.parse(xhr.responseText);
+				let rp = xhr.responseText;
+				let cinfo = JSON.parse(rp);
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+						
+					} else {
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						chrome.storage.local.set({token: ''});
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+						return false;
+					}
+				}
 
-				//if(cinfo['fs'] === '1') {
-					lastseen = cinfo['lastseen'];
-					doFullSync();
-				//} else {
-				//	loglines = logit("Info: FullSync check negative");
-				//	getPHPMarks();
-				//}
+				lastseen = cinfo['lastseen'];
+				await doFullSync();
 			}
 		}
-		//loglines = logit("Info: Start Sync check");
 		xhr.send(params);
 	});
 }
 
-function doFullSync() {
+async function doFullSync() {
 	loglines = logit("Info: Sync started.");
 	try {
-		chrome.storage.local.get(null, function(options) {
+		chrome.storage.local.get(null, async function(options) {
 			if(options['s_type'] == 'PHP') {
-				getAllPHPMarks(true);
+				await getAllPHPMarks(true);
 			}
 		});
 	} catch(error) {
@@ -936,33 +1153,43 @@ function doFullSync() {
 	}
 }
 
-function getAllPHPMarks(fs=false) {
-	chrome.storage.local.get(null, function(options) {
+function getAllPHPMarks() {
+	chrome.storage.local.get(null, async function(options) {
 		let xhr = new XMLHttpRequest();
 		let params = 'client='+options['s_uuid']+'&caction=export&type=json&s='+options['actions']['startup'];
-		xhr.open('POST', options['wdurl'] + '?t=' + Math.random(), true);
+		xhr.open('POST', options['wdurl'] + '?t=' + Math.random(), false);
 		xhr.withCredentials = true;
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-		xhr.onload = function () {
+		xhr.onload = async function () {
 			if( xhr.status != 200 ) {
 				message = chrome.i18n.getMessage("errorGetBookmarks") + xhr.status;
 				notify('error',message);
-				loglines = logit('Error: '+message);
-			}
-			else {
+				loglines = logit('Error: ' + message);
+			} else {
 				let response = xhr.responseText;
-				if(response != "false") {
-					if(abrowser == false) response = c2cm(response);
-					let PHPMarks = JSON.parse(response);
-					count = 0;
-					loglines = logit('Info: Bookmarks retrieved from server');
-					//(fs === true) ? importFull(PHPMarks):importMarks(PHPMarks);
-					importFull(PHPMarks);
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) {
+					if(xtResponse !== '0') {
+						await chrome.storage.local.set({token: xtResponse});
+					} else {
+						chrome.storage.local.set({token: ''});
+						let message = chrome.i18n.getMessage("optionsLoginError");
+						notify('error', message);
+						chrome.browserAction.setBadgeText({text: '!'});
+						chrome.browserAction.setBadgeTextColor({color: "white"});
+						chrome.browserAction.setBadgeBackgroundColor({color: "red"});
+					}
 				}
-				else {
-					loglines = logit("Error: Error when retrieving bookmarks from server for import");
-				}
+				if(abrowser == false) response = c2cm(response);
+				let PHPMarks = JSON.parse(response);
+				count = 0;
+				loglines = logit('Info: Bookmarks retrieved from server');
+				await importFull(PHPMarks);
 			}
 
 			let date = new Date(Date.now());
@@ -978,7 +1205,6 @@ async function importFull(rMarks) {
 	const lMarks = [];
 	const dMarks = new Array();
 	const uMarks = new Array();
-	const rbmCount = rMarks.length;
 
 	createLocalMarks(oMarks);
 	function createLocalMarks(nodes) {
@@ -1055,25 +1281,22 @@ async function importFull(rMarks) {
 	}
 
 	async function iMoveMark(remoteMark) {
-		if(typeof remoteMark.bmURL !== 'undefined') {
-			let localMark = (await searchBookmarkAsync({url: remoteMark.bmURL}))[0];
-			let remoteParentFolderName = '';
-			let localParentId = '';
+		let localMark = (await searchBookmarkAsync({title: remoteMark.bmTitle}))[0];
+		let remoteParentFolderName = '';
+		let localParentId = '';
 
-			if(remoteMark.bmParentID.endsWith('_____') || remoteMark.bmParentID.length === 1) {
-				localParentId = remoteMark.bmParentID;
-			} else {
-				let searchedID = remoteMark.bmParentID;
-				remoteParentFolderName = rMarks.filter(element => element.bmID == searchedID)[0].bmTitle;		
-				localParentId = (await searchBookmarkAsync({title: remoteParentFolderName}))[0].id;
-			}
-			
-			let destination = new Object();
-			destination.parentId = localParentId;
-			destination.index = parseInt(remoteMark.bmIndex);
-
-			let newMark = (await moveBookmarkAsync(localMark.id, destination));
+		if(remoteMark.bmParentID.endsWith('_____') || remoteMark.bmParentID.length === 1) {
+			localParentId = remoteMark.bmParentID;
+		} else {
+			let searchedID = remoteMark.bmParentID;
+			remoteParentFolderName = rMarks.filter(element => element.bmID == searchedID)[0].bmTitle;		
+			localParentId = (await searchBookmarkAsync({title: remoteParentFolderName}))[0].id;
 		}
+
+		let destination = new Object();
+		destination.parentId = localParentId;
+		destination.index = parseInt(remoteMark.bmIndex);
+		let newMark = (await moveBookmarkAsync(localMark.id, destination));
 	}
 
 	chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
@@ -1122,7 +1345,7 @@ async function importFull(rMarks) {
 	chrome.bookmarks.onMoved.addListener(onMovedCheck);
 	chrome.bookmarks.onRemoved.addListener(onRemovedCheck);
 
-	exportPHPMarks(uMarks);
+	if(uMarks.length > 0) exportPHPMarks(uMarks);
 }
 
 function c2cm(bookmarks) {
@@ -1327,8 +1550,7 @@ function getDAVMarks() {
 		xhr.open('GET', options['wdurl'] + '/' + filename + '?t=' + Math.random(), true);
 		xhr.withCredentials = true;
 		xhr.setRequestHeader('X-Filename', filename);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
-		
+		xhr.setRequestHeader('Authorization', 'Basic ' + btoa(options['creds']));
 		xhr.onload = function () {		
 			if( xhr.status != 200 ) {
 				message = chrome.i18n.getMessage("errorGetBookmarks") + xhr.status;