+++ scripts/options.js
@@ -2,65 +2,54 @@ var background_page = chrome.extension.getBackgroundPage();
 
 function checkForm() {
 	var wchanged = false;
-	var uchanged = false;
-	var pchanged = false;
 
+	if((document.getElementById('wdurl').value != '') && (document.getElementById('wdurl').value != document.getElementById('wdurl').defaultValue)) {
+		document.getElementById("lginl").style.visibility = 'visible';
+	}
 
-	if(document.getElementById('wdurl').value != '' && document.getElementById('user').value != '' && document.getElementById('password').value != '' && document.querySelector('input[name="stype"]:checked').value !== true){
+	if(document.getElementById('wdurl').value != '' && document.querySelector('input[name="stype"]:checked').value !== true){
 		document.getElementById('mdownload').disabled=false;
 		document.getElementById('mupload').disabled=false;
-		document.getElementById('mremove').disabled=false;
     } else{
 		document.getElementById('mdownload').disabled=true;
 		document.getElementById('mupload').disabled=true;
-		document.getElementById('mremove').disabled=false;
 	}
 
 	if(document.getElementById('wdurl').value != document.getElementById('wdurl').defaultValue) wchanged = true;
-	if(document.getElementById('user').value != document.getElementById('user').defaultValue) uchanged = true;
-	if(document.getElementById('password').value != document.getElementById('password').defaultValue) pchanged = true;
-
-
-
-	if(wchanged && uchanged && pchanged) saveOptions();
+	if(wchanged) saveOptions();
 }
 
-function saveOptions(e) {
-	if(typeof e !== "undefined") e.preventDefault();
+function checkForm2() {
+	if(document.getElementById('nuser').value != '' && document.getElementById('npassword').value != '' ) {
+		document.getElementById('lgin').disabled = false;
+	} else {
+		document.getElementById('lgin').disabled = true;
+	}
+}
+
+function gToken(e) {
+	e.preventDefault();
+	document.getElementById('lginl').classList.add('loading');
+	document.getElementById('crdialog').style.display = "none";
+	let xhr = new XMLHttpRequest();
 	let wmessage = document.getElementById('wmessage');
-	if(typeof last_sync === "undefined" || last_sync.toString().length <= 0) {
-		document.getElementById("wmessage").textContent = chrome.i18n.getMessage("optionsNotUsed");
-		wmessage.style.cssText = "border-color: darkorange; background-color: gold;";
+	cdata = "caction=tl&client=" + document.getElementById('s_uuid').value + "&s=" + document.getElementById('s_startup').checked;
+	let rnd = Math.floor((Math.random() * 100) + 1) + '.txt';
+	var url = document.getElementById('wdurl').value;	
+
+	if(document.getElementById('php_webdav').checked) {
+		xhr.open("POST", url, true);
+	} else {
+		url = url.endsWith('/') ? url.slice(0, -1):url;
+		url = url + "/" + rnd;
+		xhr.open("PUT", url, true);
 	}
 
-	chrome.storage.local.set({
-		actions: {
-			startup:document.querySelector("#s_startup").checked,
-			create:document.querySelector("#s_create").checked,
-			change:document.querySelector("#s_change").checked,
-			remove:document.querySelector("#s_remove").checked,
-			crsrv:document.querySelector("#b_action").checked
-		},
+	var creds = btoa(document.getElementById('nuser').value + ':' + document.getElementById('npassword').value);
 
-		s_type: document.querySelector('input[name="stype"]:checked').value,
-		s_uuid: document.querySelector("#s_uuid").value,
-		wdurl: document.querySelector("#wdurl").value,
-
-		creds: btoa(document.querySelector("#user").value+':'+document.querySelector("#password").value)
-	});
-
-	let xhr = new XMLHttpRequest();
-	let cdata = "caction=logout&client="+document.getElementById('s_uuid').value;
-	xhr.open("POST", document.getElementById('wdurl').value, true);
-	xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-	xhr.send(cdata);
-
-	cdata = "client=" + document.getElementById('s_uuid').value + "&caction=tl&s="+document.getElementById('s_startup').checked;
-	xhr.open("POST", document.getElementById('wdurl').value, true);
-	xhr.setRequestHeader('Authorization', 'Basic ' + btoa(document.getElementById('user').value + ':' + document.getElementById('password').value));
+	xhr.setRequestHeader('Authorization', 'Basic ' + creds);
 	xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 	xhr.withCredentials = true;
-	
 	xhr.onload = function () {
 		switch(xhr.status) {
 			case 404: 	wmessage.textContent = chrome.i18n.getMessage("optionsErrorURL");
@@ -71,29 +60,93 @@ function saveOptions(e) {
 						wmessage.style.cssText = "border-color: red; background-color: lightsalmon;";
 						console.error('Syncmarks Error: '+chrome.i18n.getMessage("optionsErrorUser"));
 						break;
-			case 200:	if(xhr.responseText.indexOf('updated') == 7) {
-							wmessage.textContent = chrome.i18n.getMessage("optionsSuccessLogin");
-							wmessage.style.cssText = "border-color: green; background-color: #98FB98;";
+			case 200:	let rp = xhr.responseText;
+						let response = (rp.indexOf('<') === -1) ? JSON.parse(rp):'0';
+						if(typeof response.length === 'undefined' && response.token.length != 0) {
+							document.getElementById('lginl').classList.remove('loading');
+							document.getElementById('lginl').style.visibility = "hidden";
 							chrome.storage.local.set({
 								wdurl: document.getElementById('wdurl').value,
-								creds: btoa(document.querySelector("#user").value+':'+document.querySelector("#password").value)
+								token: response.token,
+								s_type: 'PHP',
 							});
-						} else {
-							wmessage.textContent = 'Warning: '+xhr.responseText;
-							wmessage.style.cssText = "border-color: red; background-color: lightsalmon;";
-							console.warn('Syncmarks Warning: '+xhr.responseText);
+							chrome.storage.local.remove("creds");
 						}
+
+						if(response.message.indexOf('updated') == 7) {
+							wmessage.textContent = chrome.i18n.getMessage("optionsSuccessLogin");
+							wmessage.style.cssText = "border-color: green; background-color: #98FB98;";							
+						} else {
+							wmessage.textContent = 'Warning: '+ response.message;
+							wmessage.style.cssText = "border-color: red; background-color: lightsalmon;";
+							console.warn('Syncmarks Warning: '+ response.message);
+						}
+						
 						break;
+			case 201:	xhr.open("DELETE", url, true);
+						xhr.onload = function () {
+							let status = this.status;
+							if(status >= 200 || status < 300) {
+								document.getElementById('lginl').classList.remove('loading');
+								document.getElementById('lginl').style.visibility = "hidden";
+								chrome.storage.local.set({
+									wdurl: document.getElementById('wdurl').value,
+									creds: creds,
+									s_type: 'WebDAV',
+								});
+								chrome.storage.local.remove("token");
+							} else {
+								document.getElementById('lginl').classList.remove('loading');
+								chrome.storage.local.set({
+									wdurl: document.getElementById('wdurl').value,
+									creds: '',
+									s_type: 'WebDAV',
+								});
+								chrome.storage.local.remove("token");
+							}
+						}
+						xhr.send(cdata);
+						break;			
 			default:	wmessage.textContent = chrome.i18n.getMessage("optionsErrorLogin") + xhr.status;
 						wmessage.style.cssText = "background-color: #ff7d52;";
 						console.error('Syncmarks Error: '+chrome.i18n.getMessage("optionsErrorLogin") + xhr.status);
 						break;
 		}
 		wmessage.className = "show";
-		setTimeout(function(){wmessage.className = wmessage.className.replace("show", ""); }, 3000);
+		setTimeout(function(){wmessage.className = wmessage.className.replace("show", "hide"); }, 3000);
 		background_page.init();
 	};
 	xhr.send(cdata);
+}
+
+function saveOptions(e) {
+	if(typeof e !== "undefined") e.preventDefault();
+	var wmessage = document.getElementById('wmessage');
+	wmessage.textContent = chrome.i18n.getMessage("optionsSuccessLogin");
+
+	if(typeof last_sync === "undefined" || last_sync.toString().length <= 0) {
+		document.getElementById("wmessage").textContent = chrome.i18n.getMessage("optionsNotUsed");
+		wmessage.style.cssText = "border-color: darkorange; background-color: gold;";
+	}
+
+	chrome.storage.local.set({
+		actions: {
+			startup:document.getElementById("s_startup").checked,
+			create:document.getElementById("s_create").checked,
+			change:document.getElementById("s_change").checked,
+			remove:document.getElementById("s_remove").checked,
+			crsrv:document.getElementById("b_action").checked
+		},
+
+		s_type: document.querySelector('input[name="stype"]:checked').value,
+		s_uuid: document.getElementById("s_uuid").value,
+		wdurl: document.getElementById("wdurl").value,
+	});
+
+	wmessage.style.cssText = "border-color: green; background-color: #98FB98;";
+	wmessage.className = "show";
+	setTimeout(function(){wmessage.className = wmessage.className.replace("show", "hide"); }, 5000);
+		
 }
 
 function rName() {
@@ -102,7 +155,11 @@ function rName() {
 		var xhr = new XMLHttpRequest();
 		let cdata = "client="+options['s_uuid']+"&caction=arename&nname="+name;
 		xhr.open("POST", options['wdurl'], true);
-		xhr.setRequestHeader("Authorization", 'Basic ' + options['creds']);
+		let tarr = {};
+		tarr['client'] = options['s_uuid'];
+		tarr['token'] = options['token'];
+		if(tarr['token'] == '') return false;
+		xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 		xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 		xhr.withCredentials = true;
 		xhr.onload = function () {
@@ -110,6 +167,9 @@ function rName() {
 				message = "Error set name of client."  + xhr.status;
 				background_page.notify('error',message);
 				background_page.loglines = background_page.logit('Error: '+message);
+			} else {
+				let xtResponse = xhr.getResponseHeader("X-Request-Info");
+				if(xtResponse !== null) chrome.storage.local.set({token: xtResponse});
 			}
 		}
 		xhr.send(cdata);
@@ -120,7 +180,11 @@ function gName() {
 	let xhr = new XMLHttpRequest();
 	let cdata = "client=" + document.getElementById("s_uuid").value + "&caction=cinfo";
 	xhr.open("POST", document.getElementById("wdurl").value, true);
-	xhr.setRequestHeader('Authorization', 'Basic ' + btoa(document.getElementById('user').value + ':' + document.getElementById('password').value));
+	let tarr = {};
+	tarr['client'] = options['s_uuid'];
+	tarr['token'] = options['token'];
+	if(tarr['token'] == '') return false;
+	xhr.setRequestHeader('Authorization', 'Bearer ' + btoa(encodeURIComponent(JSON.stringify(tarr))));
 	xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 	xhr.withCredentials = true;
 	xhr.onload = function () {
@@ -129,9 +193,12 @@ function gName() {
 			background_page.notify('error',message);
 			background_page.loglines =  background_page.logit('Error: '+message);
 		} else {
+			let xtResponse = xhr.getResponseHeader("X-Request-Info");
+			if(xtResponse !== null) chrome.storage.local.set({token: xtResponse});
+
 			let response = JSON.parse(xhr.responseText);		
 			document.getElementById("cname").title = (response) ? document.getElementById('s_uuid').value + " (" + response.ctype + ")":document.getElementById('s_uuid').value;
-			document.getElementById("cname").defaultValue = (response.cname == document.getElementById('s_uuid').value) ? '':response.cname;
+			if(response !== null) document.getElementById("cname").defaultValue = (response.cname == document.getElementById('s_uuid').value) ? '':response.cname;
 		}
 	}
 	xhr.send(cdata);
@@ -146,13 +213,8 @@ function restoreOptions() {
 			wmessage.className = "show";
 			setTimeout(function(){wmessage.className = wmessage.className.replace("show", ""); }, 3000);
 		}
-		document.querySelector("#wdurl").defaultValue = options['wdurl'] || "";
-		if(options['creds'] !== undefined) {
-			let creds = atob(options['creds']).split(':');
-			document.querySelector("#user").defaultValue = creds[0] || "";
-			document.querySelector("#password").defaultValue = creds[1] || "";
-		}
-
+		document.querySelector("#wdurl").defaultValue = options['wdurl'] || "";		
+		
 		if(options['s_uuid'] === undefined) {
 			let nuuid = background_page.uuidv4();
 			document.getElementById("s_uuid").defaultValue = nuuid;
@@ -173,17 +235,28 @@ function restoreOptions() {
 		} else {
 			document.getElementById("s_auto").defaultChecked = false;
 		}
-
-		if("s_type" in options) {
+		
+		if("s_type" in options) {		
 			document.querySelector('input[name="stype"][value="'+ options['s_type'] +'"]').defaultChecked = true;
 			if(options['s_type'] == 'PHP') {
 				gName();
 				document.querySelector("#cname").placeholder = document.querySelector("#s_uuid").defaultValue;
 				document.getElementById("php_webdav").checked = true;
+				if(options['token'] === '') document.getElementById("lginl").style.visibility = 'visible';
 			} else {
 				document.getElementById("php_webdav").checked = false;
+				if(options['creds'] === '') document.getElementById("lginl").style.visibility = 'visible';
 			}
+			
 		}
+
+		chrome.commands.getAll((commands) => {
+			for (let {name, shortcut} of commands) {
+				var s = (name === 'bookmark-tab') ? shortcut:'undef';
+			}
+
+			document.getElementById("obmd").innerText = document.getElementById("obmd").innerText + ` (${s})`;
+		});
 
 		last_sync = options['last_sync'] || 0;
 		if(last_sync.toString().length > 0) {
@@ -196,8 +269,10 @@ function restoreOptions() {
 	});
 }
 
-function exportOptions() {
+function exportOptions(e) {
+	e.preventDefault();
 	chrome.storage.local.get(null, function(options) {
+		delete options.clist;
 		let confJSON = JSON.stringify(options);
 		let dString = new Date().toISOString().slice(0,10);
 		let nameStr = document.getElementById('cname').value.replace(/[^a-z0-9]/gi, '_').toLowerCase();
@@ -208,6 +283,7 @@ function exportOptions() {
 		document.body.appendChild(element);
 		element.click();
 		document.body.removeChild(element);
+		document.getElementById('expimpdialog').style.display = 'none';
 	});
 }
 
@@ -227,28 +303,25 @@ function importOptions() {
 			},
 			s_type: ioptions.s_type,
 			wdurl: ioptions.wdurl,
-			creds: ioptions.creds,
 		});
 
 		if(resCID) {
+			document.querySelector("#cname").placeholder = ioptions.s_uuid;
+			document.querySelector("#s_uuid").value = ioptions.s_uuid;
+			document.getElementById("lginl").style.visibility = 'hidden';
 			chrome.storage.local.set({
-				s_uuid: ioptions.s_uuid
+				s_uuid: ioptions.s_uuid,
+				token: ioptions.token,
+				creds: ioptions.creds,
 			});
 		}
-		
-		let creds = atob(ioptions.creds).split(':');
 
 		document.querySelector("#wdurl").value = ioptions.wdurl;
-		document.querySelector("#user").value = creds[0];
-		document.querySelector("#password").value = creds[1];
-		if(resCID) document.querySelector("#s_uuid").value = ioptions.s_uuid;
 		document.querySelector("#s_startup").checked = ioptions.actions.startup;
 		document.querySelector("#s_create").checked = ioptions.actions.create;
 		document.querySelector("#s_change").checked = ioptions.actions.change;
 		document.querySelector("#s_remove").checked = ioptions.actions.remove;
 		document.querySelector("#b_action").checked = ioptions.actions.crsrv;
-
-		if(resCID) document.querySelector("#cname").placeholder = ioptions.s_uuid;
 
 		wmessage.textContent = chrome.i18n.getMessage("optionsSuccessImport");
 		wmessage.style.cssText = "border-color: green; background-color: #98FB98;";
@@ -264,16 +337,11 @@ function importOptions() {
 
 function manualImport(e) {
 	e.preventDefault();
-	let fs = true;
-	if (this.id === 'iyes') {
-		background_page.removeAllMarks();
-		fs = false;
-	}
+	if (this.id === 'iyes') background_page.removeAllMarks();
 
 	try {
 		chrome.storage.local.get(null, function(options) {
 			if(options['s_type'] == 'PHP') {
-				//background_page.getAllPHPMarks(fs);
 				background_page.getAllPHPMarks(true);
 			} else if (options['s_type'] == 'WebDAV') {
 				background_page.getDAVMarks();
@@ -287,27 +355,13 @@ function manualImport(e) {
 	}
 }
 
-function manualRemove(e) {
-	e.preventDefault();
-	try {
-		background_page.removeAllMarks();
-	}
-	catch(error) {
-		background_page.loglines = background_page.logit(error);
-	}
-	finally {
-		document.getElementById("rmdialog").style.display = "none";
-		chrome.storage.local.set({last_s: 1});
-	}
-}
-
 function manualExport(e) {
 	e.preventDefault();
 	var background_page = chrome.extension.getBackgroundPage();
 	try {
 		if(document.querySelector('input[name="stype"]:checked').value == 'WebDAV') {
 			background_page.saveAllMarks();
-		} else if(document.querySelector('input[name="stype"]:checked').value == 'PHP') {
+		} else if (document.querySelector('input[name="stype"]:checked').value == 'PHP') {
 			background_page.exportPHPMarks();
 		}
 	} catch(error) {
@@ -331,6 +385,10 @@ function localizeHtmlPage() {
 			obj.innerText = valNewH;
         }
     }
+
+	document.getElementById('blbl').title = chrome.i18n.getMessage("backendType");
+	document.getElementById('oauto').title = chrome.i18n.getMessage("ManAuto");
+	document.getElementById('obmd').title = chrome.i18n.getMessage("toBackend");
 }
 
 function openTab(tabname) {
@@ -392,11 +450,11 @@ function switchBackend() {
 	if(backendPHP) {
 		document.getElementById("php").checked = true;
 		document.getElementById("wdav").checked = false;
-		document.getElementById("blbl").innerText = "Backend PHP";
+		document.getElementById("blbl").innerText = "Server: PHP";
 	} else {
 		document.getElementById("php").checked = false;
 		document.getElementById("wdav").checked = true;
-		document.getElementById("blbl").innerText = "Backend WebDAV";
+		document.getElementById("blbl").innerText = "Server: WebDAV";
 	}
 }
 
@@ -435,7 +493,6 @@ window.addEventListener('load', function () {
     rawFile.send(null);
 
 	var imodal = document.getElementById("impdialog");
-	var rmodal = document.getElementById("rmdialog");
 	var emodal = document.getElementById("expdialog");
 	var comodal = document.getElementById("expimpdialog");
 
@@ -451,21 +508,27 @@ window.addEventListener('load', function () {
 	document.getElementById("confinput").addEventListener('change', importOptions);
 	document.getElementById("iyes").addEventListener("click", manualImport);
 	document.getElementById("eyes").addEventListener("click", manualExport);
-	document.getElementById("ryes").addEventListener("click", manualRemove);
 	document.getElementById("ino").addEventListener("click", manualImport);
-	document.getElementById("rno").addEventListener("click", function() { rmodal.style.display = "none";});
 	document.getElementById("eno").addEventListener("click", function() { emodal.style.display = "none";});
 	document.getElementById("iclose").addEventListener("click", function() {imodal.style.display = "none";});
-	document.getElementById("rclose").addEventListener("click", function() {rmodal.style.display = "none";});
 	document.getElementById("eclose").addEventListener("click", function() {emodal.style.display = "none";});
+	document.getElementById("crclose").addEventListener("click", function() {document.getElementById("crdialog").style.display = "none";});
 	document.getElementById("mdownload").addEventListener("click", function() {imodal.style.display = "block"});
-	document.getElementById("mremove").addEventListener("click", function() {rmodal.style.display = "block"})
 	document.getElementById("mupload").addEventListener("click", function() {emodal.style.display = "block"});
 	document.getElementById("wdurl").addEventListener("change", checkForm);
-	document.getElementById("user").addEventListener("change", checkForm);
-	document.getElementById("password").addEventListener("change", checkForm);
+	document.getElementById("lginl").addEventListener("click", function(e) {
+		e.preventDefault;
+		document.getElementById("nuser").defaultValue = '';
+		document.getElementById("npassword").defaultValue = '';
+		document.getElementById("crdialog").style.display = "block";
+		document.getElementById("nuser").focus();
+	});
+
+	document.getElementById("nuser").addEventListener("input", checkForm2);
+	document.getElementById("npassword").addEventListener("input", checkForm2);
 	
 	document.getElementById("php_webdav").addEventListener("change", switchBackend);
+	document.getElementById('lgin').addEventListener("click", gToken);
 	
 	document.getElementById("s_startup").addEventListener("change", saveOptions);
 	document.getElementById("s_create").addEventListener("change", saveOptions);
@@ -489,11 +552,6 @@ window.addEventListener('load', function () {
 		if (event.target == imodal) {
 			imodal.style.display = "none";
 		}
-
-		if (event.target == rmodal) {
-			rmodal.style.display = "none";
-		}
-
 		if (event.target == emodal) {
 			emodal.style.display = "none";
 		}