+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ChangeLog
 =========
+v1.19.4 (2022-03-12)
+-------------------------
+- Added token based login
+- Fixed encoding issue
+- Fixed import issue
+- Fixed WebDAV login
+
+
 v1.19.3 (2022-02-25)
 -------------------------
 - Minor Fixes
@@ -7,7 +15,7 @@ v1.19.3 (2022-02-25)
 
 v1.19.2 (2022-02-24)
 -------------------------
-- Fix sync for unsynced client
+- Fix sync for un-synced client
 
 
 v1.19.1 (2022-02-21)